requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0
woocommerce>=3.0.0
pandas>=2.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
chromadb>=0.4.0
openai>=1.0.0
pydantic>=2.0.0
httpx>=0.25.0
tenacity>=8.2.0